Support an array of shell provisioners
Replaced uncessary string interpretation with bare variables

Something must have told me about that. An error? A linter? Again, by
the time I got to comitting this, I completely forgot why I made that
chage. It seems to work, so I am trusting past-self. If it's wrong,
well, someone will find it and fix it. Might not be the same person.

Free as in puppies!